Fix documentation formatting and examples to pass CI
