--- src/cypher/ast.rs
@@ -1,7 +1,0 @@
-//! Cypher AST types.
-
-/// Cypher AST query structure.
-#[derive(Debug, Clone, PartialEq)]
-pub struct CypherQuery {
-    // Placeholder
-}
