--- src/cypher/parser.rs
@@ -1,4 +1,0 @@
-//! Cypher parser.
-
-/// Recursive descent parser for Cypher queries.
-pub struct Parser;
