--- src/cypher/tests.rs
@@ -1,7 +1,0 @@
-//! Cypher module tests.
-
-#[test]
-fn test_cypher_placeholder() {
-    // Placeholder assertion to ensure test infrastructure works
-    assert_eq!(1 + 1, 2);
-}
