--- src/cypher/transform.rs
@@ -1,4 +1,0 @@
-//! Cypher AST -> QueryOp transformation.
-
-/// Transformer from Cypher AST to QueryOp IR.
-pub struct Transformer;
