+++ src/lib.rs
@@ -25,9 +25,8 @@
 //!
 //! ```rust,no_run
 //! use gallifreydb::{GallifreyDB, properties, WriteOps};
-//! use gallifreydb::core::temporal::time;
 //!
-//! # fn main() -> Result<(), Box<dyn std::error::Error>> {
+//! # fn main() -> gallifreydb::Result<()> {
 //! let db = GallifreyDB::new()?;
 //!
 //! // Create a node
@@ -38,17 +37,17 @@
 //!
 //! # let timestamp = gallifreydb::core::temporal::time::now();
 //! // Later, update a property
-//! db.write(|tx| tx.update_node(alice, properties! {
-//!     "age" => 31,
-//! }))?;
+//! db.write(|tx| {
+//!     tx.update_node(alice_id, properties! {
+//!         "age" => 31,
+//!     })
+//! })?;
 //!
 //! // Query current state
 //! let current = db.get_node(alice_id)?;
 //!
 //! // Time-travel to see historical state
-//! // Use current time as a placeholder for the point in time we want to query
-//! let now = time::now();
-//! let historical = db.get_node_at_time(alice, now, now)?;
+//! let historical = db.get_node_at_time(alice_id, timestamp, timestamp)?;
 //! # Ok(())
 //! # }
 //! ```
@@ -79,9 +78,6 @@ pub mod honeycomb;
 // Optional SQL:2011 temporal syntax support
 #[cfg(feature = "sql")]
 pub mod sql;
-// Optional Cypher Query Language support
-#[cfg(feature = "cypher")]
-pub mod cypher;
 // Optional HTTP server module
 #[cfg(feature = "http-server")]
 pub mod http;
